Move application delegate functionality out of Objective-C methods
